Consolidate unit and load tests
