+++ .bumpversion.cfg
@@ -27,7 +27,7 @@ first_value = 1
 
 [bumpversion:file:CONTRIBUTING_COMMON_ERRORS.md]
 
-[bumpversion:file:tests/test_metadata.py]
+[bumpversion:file:tests/unit/test_metadata.py]
 
 [bumpversion:file:docs/source/install.rst]
