+++ tests/unit/test_athena.py
@@ -9,7 +9,7 @@ import pytest
 
 import awswrangler as wr
 
-from ._utils import (
+from .._utils import (
     ensure_athena_ctas_table,
     ensure_athena_query_metadata,
     ensure_data_types,
