+++ tests/unit/test_athena_cache.py
@@ -6,7 +6,7 @@ import pytest
 
 import awswrangler as wr
 
-from ._utils import ensure_athena_query_metadata
+from .._utils import ensure_athena_query_metadata
 
 logging.getLogger("awswrangler").setLevel(logging.DEBUG)
