+++ tests/unit/test_mysql.py
@@ -9,7 +9,7 @@ from pymysql.cursors import SSCursor
 
 import awswrangler as wr
 
-from ._utils import ensure_data_types, get_df
+from .._utils import ensure_data_types, get_df
 
 logging.getLogger("awswrangler").setLevel(logging.DEBUG)
