+++ tests/unit/test_neptune.py
@@ -11,7 +11,7 @@ from gremlin_python.process.traversal import Direction, T
 
 import awswrangler as wr
 
-from ._utils import extract_cloudformation_outputs
+from .._utils import extract_cloudformation_outputs
 
 logging.getLogger("awswrangler").setLevel(logging.DEBUG)
