+++ tests/unit/test_session.py
