+++ tests/unit/test_sqlserver.py
@@ -9,7 +9,7 @@ import pytest
 
 import awswrangler as wr
 
-from ._utils import ensure_data_types, get_df
+from .._utils import ensure_data_types, get_df
 
 logging.getLogger("awswrangler").setLevel(logging.DEBUG)
