+++ tox.ini
@@ -16,7 +16,7 @@ deps =
        s3fs==0.4.2
 commands =
        pytest -n {posargs} -s -v --timeout=300 --reruns=2 --reruns-delay=15 \
-              --junitxml=test-reports/junit.xml --log-file=test-reports/logs.txt tests
+              --junitxml=test-reports/junit.xml --log-file=test-reports/logs.txt tests/unit
 
 [testenv:py{37,38,39,310}-distributed]
 passenv = AWS_PROFILE AWS_DEFAULT_REGION AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY AWS_SESSION_TOKEN
@@ -33,4 +33,4 @@ deps =
        pytest-cov==3.0.0
 commands =
        pytest -n {posargs} -s -v --timeout=300 --reruns=2 --reruns-delay=60 --cov=awswrangler \
-              --junitxml=test-reports/junit.xml --log-file=test-reports/logs.txt tests
+              --junitxml=test-reports/junit.xml --log-file=test-reports/logs.txt tests/unit
